House Keeping (again)! - Speeding up the stored procs - Simplifying the stored procs - SImplifying and speeding up some in-line SQL - Adding transactions to in-line SQL with temp tables

+++ src/CardboardBox.Anime.Database/Models/Manga/MangaProgress.cs
@@ -35,7 +35,7 @@ public class MangaStats
 	public int MaxChapterNum { get; set; }
 
 	[JsonPropertyName("chapterNum")]
-	public int ChapterNum { get; set; }
+	public double ChapterNum { get; set; }
 
 	[JsonPropertyName("pageCount")]
 	public int PageCount { get; set; }
